`Http2SettingsBuilder.initialWindowSize` should accept `int`, not `long` (#2573)
Motivation:

RFC says: Values above the maximum flow-control window size of 2^31-1
MUST be treated as a connection error of type `FLOW_CONTROL_ERROR`.
To prevent users from accidentally setting incorrect values, we should
accept only `int`. Also, `Http2Settings.initialWindowSize()` return type
is `Integer`. Builder should be consistent.

Modifications:

- Deprecate `Http2SettingsBuilder.initialWindowSize(long)`;
- Add `Http2SettingsBuilder.initialWindowSize(int)`;
- Change `validate32Unsigned` and `validate31Unsigned` to throw an
exception only when the passed value is greater than the max value;
- Remove unnecessary type cast for `maxHeaderListSize` in
`H2ProtocolConfigBuilder`;

Result:

1. Users can not accidentally configure too large `initialWindowSize`
value.
2. `Integer.MAX_VALUE` and `MAX_UNSIGNED_INT` are accepted values for
settings.

+++ servicetalk-http-netty/src/main/java/io/servicetalk/http/netty/H2ProtocolConfigBuilder.java
@@ -54,7 +54,7 @@ public final class H2ProtocolConfigBuilder {
     private static final int DEFAULT_FLOW_CONTROL_QUANTUM = 1024 * 16;
     private Http2Settings h2Settings = new Http2SettingsBuilder()
             .initialWindowSize(INITIAL_FLOW_CONTROL_WINDOW)
-            .maxHeaderListSize((int) DEFAULT_HEADER_LIST_SIZE)
+            .maxHeaderListSize(DEFAULT_HEADER_LIST_SIZE)
             .build();
     private HttpHeadersFactory headersFactory = H2HeadersFactory.INSTANCE;
     private BiPredicate<CharSequence, CharSequence> headersSensitivityDetector = DEFAULT_SENSITIVITY_DETECTOR;
